refactor(services): replace quota-warnings notifier with QuotaWarningNotifier
- New Pi-agnostic QuotaWarningNotifier class in services/
- Accepts NotifyFn callback instead of Pi UI dependencies
- Escalation/cooldown logic fully testable
- Remove old notifier.ts and notifier.test.ts from quota-warnings extension
